requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
boto3==1.17.1
botocore==1.20.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-appconf==1.0.4
django-better-admin-arrayfield @ git+https://github.com/robert8888/django-better-admin-arrayfield.git@769532aceec0292d6c46facb91ea9a319296d4fe
django-ckeditor==6.0.0
django-compressor==2.4
django-compressor-toolkit==0.6.0
django-fieldsets-with-inlines==0.6
django-heroku==0.3.1
django-imagefield==0.13.0
django-js-asset==1.2.2
django-libsass==0.8
django-modeladmin-reorder==0.3.1
django-nested-admin==3.3.3
django-parler==2.2
django-polymorphic==3.0.0
django-prettyjson==0.4.1
django-resized==0.3.11
django-rosetta==0.9.5
django-solo==1.1.5
django-storages==1.11.1
django-webp-converter==0.3.0
django-webpack-loader==0.6.0
emoji==1.2.0
exrex==0.10.5
filelock==3.0.12
ftfy==5.9
gunicorn==20.0.4
httpagentparser==1.9.1
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
libsass==0.20.1
memoization==0.3.2
pbr==5.5.1
Pillow==8.1.0
polib==1.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydash==4.9.3
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-magic==0.4.22
python-monkey-business==1.0.0
pytz==2020.5
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
s3transfer==0.3.4
six==1.15.0
sqlescapy==1.0.1
sqlparse==0.4.1
standardjson==0.3.1
stevedore==3.3.0
urllib3==1.26.3
virtualenv==20.4.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
whitenoise==5.2.0
